+++ tests/test_tebd.py
@@ -7,6 +7,7 @@ import numpy as np
 from tenpy.networks.mps import MPS
 from tenpy.models.spins import SpinChain
 import tenpy.algorithms.tebd as tebd
+from tenpy.networks.site import SpinHalfSite
 from tenpy.algorithms.exact_diag import ExactDiag
 from nose.plugins.attrib import attr
 
@@ -92,7 +93,33 @@ def test_tebd():
         yield check_tebd, bc_MPS
 
 
+def test_RandomUnitaryEvolution():
+    L = 8
+    spin_half = SpinHalfSite(conserve='Sz')
+    psi = MPS.from_product_state([spin_half]*L, [0, 1]*(L//2), bc='finite')  # Neel state
+    assert tuple(psi.chi) == (1, 1, 1, 1, 1, 1, 1)
+    TEBD_params = dict(N_steps=2, trunc_params={'chi_max':10})
+    eng = tebd.RandomUnitaryEvolution(psi, TEBD_params)
+    eng.run()
+    print(eng.psi.chi)
+    assert tuple(eng.psi.chi) == (2, 4, 8, 10, 8, 4, 2)
+
+    # infinite versions
+    TEBD_params['trunc_params']['chi_max'] = 20
+    psi = MPS.from_product_state([spin_half]*2, [0, 0], bc='infinite')
+    eng = tebd.RandomUnitaryEvolution(psi, TEBD_params)
+    eng.run()
+    print(eng.psi.chi)
+    assert tuple(eng.psi.chi) == (1, 1)  # all up can not be changed
+    eng.psi = MPS.from_product_state([spin_half]*2, [0, 1], bc='infinite')
+    eng.run()
+    print(eng.psi.chi)
+    assert tuple(eng.psi.chi) == (16, 8)
+
+
+
 if __name__ == "__main__":
+    test_RandomUnitaryEvolution()
     for f_args in test_tebd():
         f = f_args[0]
         print("=" * 80)