+++ CHANGELOG.rst
@@ -44,6 +44,7 @@ Added
 - :meth:`tenpy.models.model.NearestNeighborModel.calc_H_MPO_from_bond` and
   :meth:`tenpy.models.model.MPOModel.calc_H_bond_from_MPO` for conversion of H_bond into H_MPO and vice
   versa.
+- :class:`tenpy.algorithms.tebd.RandomUnitaryEvolution` for random unitary circuits
 
 Changed
 ^^^^^^^